# Conflicts:
#	src/Main.java
#	src/Utilities/Constants/GridSize.java

+++ src/Entities/Damage/DamageFactory.java
@@ -5,7 +5,7 @@ import org.jetbrains.annotations.NotNull;
 
 public class DamageFactory {
     @Contract(value = "_, _ -> new", pure = true)
-    public static @NotNull IDamage createDamage(DamageType damageType, int value){
+    public static @NotNull IDamage createDamage(DamageType damageType, Double value){
         return new Damage(damageType, value);
     }
 }
